api/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-environ==0.10.0
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==18.4.0
freezegun==1.2.2
mysqlclient==2.1.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
